Milestone 3
Site is coming around fine.  Assets and content added.  Most of the JS will be concentrated on in the past-races section and I figured out I could toggle the links onto the section.  Ill need to figure out if I can create an array with the IDs assigned to each "blog" post of my run.

Also missing is trying to figure out the Instagram Embed.  I want to have an embed on the gallery page of my instagram running profile, not just embed one post.  Will try to figure out that next milestone.

The spotify will now be a nice to have.  I know where I want the player to be at but I'm prioritizing the bulk of the content and getting them to present well using JS.  Looking up the spotify documentation is a distraction at the moment and I want to focus on the project on hand that encompasses what I learned in class.